// math-common/build/notationTypes.d.ts
export interface RectAttributes {
    fromCol: number;
    toCol: number;
    fromRow: number;
    toRow: number;
    value: string;
}
export interface LineAttributes {
    fromCol: number;
    toCol: number;
    row: number;
}
export interface PointAttributes {
    col: number;
    row: number;
    value: string;
}